fix: address PR #9 review comments
- Fix critical JSONL parsing error in bulk-search-cards.ts
  - Changed from parsing entire stdout as single JSON to line-by-line parsing
  - Prevents failures when multiple search results are returned

- Update documentation for ygo_bulk_search
  - Removed incorrect 'not yet implemented as CLI' note from SHELL_COMMANDS.md

- Refactor duplicate project root finding logic
  - Created src/utils/project-root.ts utility function
  - Updated search-cards.ts and ygo-seek.ts to use common utility

Addresses Gemini Code Assist review feedback on commit 81d5736
